requirements.txt
flask==3.0.3
gunicorn==23.0.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
nltk==3.9.1
tensorflow-cpu==2.17.0
